feat: refactor to a Navbar component
